base testing for bodyHanddetection
