+++ components/detection/BodyHandJointsDetector/src/CommonBehavior.ice
@@ -1,0 +1,34 @@
+//******************************************************************
+// 
+//  Generated by RoboCompDSL
+//  
+//  File name: CommonBehavior.ice
+//  Source: CommonBehavior.idsl
+//
+//******************************************************************
+#ifndef ROBOCOMPCOMMONBEHAVIOR_ICE
+#define ROBOCOMPCOMMONBEHAVIOR_ICE
+module RoboCompCommonBehavior
+{
+	enum State {  Starting, Running };
+	struct Parameter
+	{
+		bool editable;
+		string value;
+		string type;
+	};
+	dictionary <string,Parameter> ParameterList;
+	interface CommonBehavior
+	{
+		ParameterList getParameterList ();
+		idempotent int getPeriod ();
+		State getState ();
+		void killYourSelf ();
+		void reloadConfig ();
+		void setParameterList (ParameterList l);
+		void setPeriod (int period);
+		idempotent int timeAwake ();
+	};
+};
+
+#endif
