--- components/detection/bodyhandjointsdetector/BodyHandJointsDetector.cdsl
@@ -1,10 +1,0 @@
-import "BodyHandJointsDetector.idsl";
-
-Component BodyHandJointsDetector
-{
-	Communications
-	{
-		implements BodyHandJointsDetector;
-	};
-	language python;
-};
