--- components/detection/bodyhandjointsdetector/etc/config
@@ -1,14 +1,0 @@
-# Endpoints for implements interfaces
-BodyHandJointsDetector.Endpoints=tcp -p 11000
-
-
-
-
-
-Ice.Warn.Connections=0
-Ice.Trace.Network=0
-Ice.Trace.Protocol=0
-Ice.MessageSizeMax=20004800
-
-
-
