--- components/detection/bodyhandjointsdetector/src/CMakeLists.txt
@@ -1,6 +1,0 @@
-cmake_minimum_required(VERSION 2.6)
-PROJECT(  BodyHandJointsDetector )
-
-INCLUDE( /opt/robocomp/cmake/robocomp.cmake )
-
-ROBOCOMP_IDSL_TO_ICE( CommonBehavior BodyHandJointsDetector )
