--- components/detection/bodyhandjointsdetector/src/inference/constant.py
@@ -1,13 +1,0 @@
-AVG_PERSON_HEIGHT = 180
-FIX_SIZE = 320
-
-# pelvis (body center) is missing, id == 2
-map_id_to_panoptic = [1, 0, 9, 10, 11, 3, 4, 5, 12, 13, 14, 6, 7, 8, 15, 16, 17, 18]
-
-limbs = [[18, 17, 1],
-         [16, 15, 1],
-         [5, 4, 3],
-         [8, 7, 6],
-         [11, 10, 9],
-         [14, 13, 12]]
-
