+++ components/detection/test/bodyHandJointsDetectorClient/etc/config
@@ -1,6 +1,6 @@
 # Proxies for required interfaces
-BodyHandJointsDetectorProxy = bodyhandjointsdetector:tcp -h localhost -p 0
-CameraSimpleProxy = camerasimple:tcp -h localhost -p 0
+BodyHandJointsDetectorProxy = bodyhandjointsdetector:tcp -h localhost -p 11000
+CameraSimpleProxy = camerasimple:tcp -h localhost -p 10005
