+++ components/detection/test/bodyHandJointsDetectorClient/src/BodyHandJointsDetector.ice
@@ -1,0 +1,34 @@
+//******************************************************************
+// 
+//  Generated by RoboCompDSL
+//  
+//  File name: BodyHandJointsDetector.ice
+//  Source: BodyHandJointsDetector.idsl
+//
+//******************************************************************
+#ifndef ROBOCOMPBODYHANDJOINTSDETECTOR_ICE
+#define ROBOCOMPBODYHANDJOINTSDETECTOR_ICE
+module RoboCompBodyHandJointsDetector
+{
+	sequence <byte> ImgType;
+	struct TImage
+	{
+		int width;
+		int height;
+		int depth;
+		ImgType image;
+	};
+	sequence <float> Points;
+	struct TBody
+	{
+		float score;
+		Points keyPoints;
+	};
+	sequence <TBody> ListFullBody;
+	interface BodyHandJointsDetector
+	{
+		ListFullBody getBodyAndHand (TImage image);
+	};
+};
+
+#endif
